# PERF: ~40x speedup in sparse init and ops by using numpy in check_integrity