@jreback 2019-01-28
seems fine for 0.24.1, @TomAugspurger ?