@TomAugspurger 2019-01-29
Since this has been around for so long, I'd lean towards 0.25.0.