@jreback 2019-01-29
@qwhelan can you move the whatsnew to 0.25.0 ping on green.